Create New Query - Final Selections
This is the monster query that will only run a few times I think against the local db. I am thinking that I will print it on a huge roll and have as part of the bed installation. 

Have tried to be as helpful as possible but this really is big and if there are examples I can follow, like multiple 'Friends' of one person then I can do more work. Or even write all the individual bits for you to combine. 

Thank you